HTTPRoute should not accept parent Services without clusterIPs (#10672)
The provisional [xRoutes GEP](https://gateway-api.sigs.k8s.io/geps/gep-1426/#allowed-service-types) says that implementations should support ClusterIP Service types as a parentRef (with or without selectors) but should not support headless services, since they do not have a "frontend" functionality (e.g an IP or DNS hostname that consumers may use). It is unclear to me whether this extends to backendRefs, but I would assume so (since there are no endpoints to send to).

Additionally, ExternalName Service types should not allowed at all, either as backendRefs or parentRefs since they pose a security threat. Last but not least LoadBalancer and NodePort types are apparently supported since they still provision a cluster IP.

We update the policy-controller to only accept a parent Service if it has a ClusterIP.

Signed-off-by: Alex Leong <alex@buoyant.io>