Explicitly use bash when using pipes in CI workflows
As detailed in the [docs](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idstepsshell), not expliciting the `shell` field results in using `bash -e`, whereas expliciting it results in `bash --noprofile --norc -eo pipefail`. We'd like to have the pipefail option set whenever pipes are used. This avoids for example an issue we had recently in the release workflow where the k3d install script download was silently failing.