Add kubelet NetworkAuthentication back into linkerd-viz (#10712)
Fixes #10699

The `NetworkAuthentication/kubelet` resource was removed in https://github.com/linkerd/linkerd2/pull/10073 because Kubernetes probes no longer need to be explicitly authorized when no other HTTPRoute resources exist on the admin Server.  However, the `linkerd viz allow-scrapes` command creates HTTPRoutes on the admin Server to authorize scrapes to the `/metrics` endpoint and therefore we need to explicitly authorize probes to the `/live` and `/ready` endpoints also.  The `allow-scrapes` command does this by referencing the `kubelet` NetworkAuthentication which was removed.

Add the `kubelet` NetworkAuthentication back since it is used by the `allow-scrapes` command.

Signed-off-by: Alex Leong <alex@buoyant.io>