Pass cluster domain flag to policy controller (#10741)
Fixes #10737

The cluster domain is not passed to the policy controller as a flag.  This means that the policy controller always uses the default value of `cluster.local`.  If the cluster's domain is different from this default, the policy controller will return incorrect authorities in it's outbound policy API.

Pass the `--cluster-domain` flag.

Signed-off-by: Alex Leong <alex@buoyant.io>