policy: add end-to-end test for failure accrual (#10681)
This branch adds an end-to-end test for failure accrual. This test
exercises the entire flow for circuit breaking, including discovering
failure accrual configuration from Service annotations and the actual
behavior of the proxy. This test does *not* exercise probation and
recovery. Instead, it asserts that once an endpoint's circuit breaker
trips, no subsequent requests are sent to that endpoint.

This test is implemented using `bb` to create a `Service` backed by an
endpoints, which always returns`500 Internal Server Error`. New test
support code is added for constructing `bb-terminus` Pods and Services.
The client is an injected `curl` pod. The existing test support code for
running `curl` doesn't quite work for this test, as the current
`curl::Runner::run` function creates a `curl` pod that executes a
_single_ `curl` command and exits. Since we need to test the client's
failure accrual behavior (which requires the _same_ client to issue
multiple requests), we add a new `curl::Runner::run_execable` method,
which creates a single `curl` pod that doesn't actually make any HTTP
requests and instead just sleeps forever. A handle to that pod is then
used to `kubectl exec` multiple `curl` commands inside the `curl`
container. We inject the `curl` pod with the proxy, annotate the
`bb-terminus` Service with failure accrual annotations (with a limit of
5 consecutive failures), and then execute 10 requests to `bb-terminus`
in the `curl` container. We assert that only the first 5 of the 10
requests fail with 500s from the `bb-terminus` backend, and that the
remainder fail with 503/504s due to the proxy's load balancer having no
ready endpoints. This ensures that the circuit breaker prevents
subsequent requests from being sent to the failing endpoint.