charts: set defaultRegistry for namespace metadata (#10739)
The namespace metadata service was not using the default registry, which was causing inconsistency with the rest of the services in the project.

Updated the namespace metadata service to use the default registry, making it consistent with the other services. 

Tested the namespace metadata service locally and verified that it is now using the default registry. Confirmed that the changes did not introduce any regressions or conflicts with other charts.

Signed-off-by: Ryan Hristovski <ryan.hristovski@docker.com>