represent TCP ports as `uint16`s
TCP ports are [defined as 16-bit unsigned integers][1]. Currently,
however, the Go controllers often represent them as other numeric types,
such as `int` and `uint32`. We currently validate parsed ports by testing
whether they are less than 65536. However, it's not always clear whether
a given `int` represents a port number that has already been validated
by this check or not.

This commit changes the `PortRange` type and parsing code to represent
ports as `uint16`, making invalid states harder to represent at the type
level. We change port parsing from using `strconv.Atoi` (which returns
an `int`, which may be negative or greater than 65535) to using
`strconv.ParseUint` with a bit size of 16 --- this function will fail to
parse the value if it is a negative number or greater than 65535, so we
no longer have to perform additional validation when parsing ports. We
can also remove some casting that is no longer necessary. It's still
necessary to cast to `uint32` in a couple of places when constructing
protobuf messages, because protobuf doesn't support 16-bit integers.

Depends on #10827

[1]: https://en.wikipedia.org/wiki/Port_(computer_networking)#Port_number