Add integration tests for loopback routing
Traffic may be routed over the loopback interface for a pod when the pod
either tries to communicate with itself using its IP, or when a pod
communicates with itself using its logical address. In the latter case,
a proportion of the traffic may be resolved to the pod's own IP by the
balancer, in which case the traffic is again routed over loopback.

This change adds an integration test to assert that locally routed
traffic does not result in any unexpected errors.

Signed-off-by: Matei David <matei@buoyant.io>