Force MeshTLSAuthentication CRs to provide at least one identity/identityRef
Fixes #10782

Added the `minItems: 1` field to `spec.identities` and
`spec.identitiRefs`. This is a BC change so it's not required to bump
the CRD version, plus current CRs not abiding to this constraint would
be broken anyway.

```bash
$ cat << EOF | k apply -f -
> apiVersion: policy.linkerd.io/v1alpha1
kind: MeshTLSAuthentication
metadata:
  name: "test"
spec:
  identities: []
> EOF
The MeshTLSAuthentication "test" is invalid: spec.identities: Invalid value: 0: spec.identities in body should have at least 1 items
```