proxy: v2.203.0
The gRPC protocol always sets the HTTP response status code to 200 and
instead communicates failures in a grpc-status header sent in a TRAILERS
frame. Linkerd uses the HTTP response status code to determine if a
response is successful, and therefore will consider all gRPC responses
successful regardless of their gRPC status code. This means that
functionality such as retries and circuit breaking do not function
correctly with gRPC traffic.

This proxy release updates the Http classifier to look for the presence
of a `Content-Type: application/grpc` header and use Grpc response
classification when it is set.

---

* Classify grpc requests properly in the http classifier (linkerd/linkerd2-proxy#2410)

Signed-off-by: Eliza Weisman <eliza@buoyant.io>