Add docker builder option to improve build tooling (#10935)
This adds the ability to pass in a docker builder option to docker. This makes building multi-arch images super simple by using our k8s infrastructure.

It also makes building multi-arch images very fast since they can be built in parallel and on native hardware.

DCO Sign off

I agree to the DCO for all the commits in this PR.

Co-authored-by: Alejandro Pedraza <alejandro.pedraza@gmail.com>
Signed-off-by: Grégoire Bellon-Gervais <gregoire.bellon-gervais@docaposte.fr>