Allow clusters to be linked without a gateway (#11226)
When a cluster has been installed without a gateway, it cannot be linked against, unless a load balancer service is used as an override. The service-mirror is tightly coupled with the notion of gateways. However, a gateway is not strictly necessary when clusters operate in a flat network.

As part of this change, `linkerd multicluster link` has been changed to allow clusters without gateways to be linked against. When a cluster does not have a gateway, all services _must_ be exported in `remote-discovery` mode, otherwise routing wouldn't work.

In addition, when a cluster does not have a gateway, linking against it will not create a probe service (since there is nothing to probe). Lastly, a check has been modified to ignore checking replicated endpoints when a service is in remote-discovery mode (to avoid false positives).

Signed-off-by: Matei David <matei@buoyant.io>