proxy: v2.208.0 (#11316)
The `linkerd-meshtls-boring` crate currently uses a Git dependency on
`boring` and `tokio-boring`. This is because, when this crate was
initially introduced, the proxy required unreleased changes to these
crates. Now, however, upstream has published all the changes we depended
on (this happened ages ago), and we can depend on these libraries from
crates.io.

This branch removes the Git deps and updates to v3.0.0 of
`boring`/`tokio-boring`. I've also changed the `cargo-deny` settings to
no longer allow Git deps on these crates, as we no longer depend on them
from Git.

---

* build(deps): use published version of `boring` (linkerd/linkerd2-proxy#2454)

Signed-off-by: Eliza Weisman <eliza@buoyant.io>
Signed-off-by: Adam Shaw <adam.shaw@vipps.no>