Update apiVersions of webhook rule (#11149)
HTTPRoute resources are sent to the validator regardless of their version (v1beta2, v1beta3) due to the match policy of the resource (resource is sent for any version of the API being served). However, it is good practice to consider all versions explicitly in the webhook rule. This change adds a wildcard operator to consider any HTTPRoute version for admission.

Signed-off-by: Takumi Sue <u630868b@alumni.osaka-u.ac.jp>