Add best-effort validation for prometheus scrape interval (#11376)
If a Linkerd-viz metrics query has a time window which is shorter than the Prometheus scrape interval, no data will be returned.  This can be confusing and unexpected.

It is difficult to validate that the time window is at least as long as the scrape interval if an external prometheus is used.  Therefore, we do a best-effort validation where we validate the scrape interval if we can find it in the prometheus config in the default location.

Signed-off-by: Alex Leong <alex@buoyant.io>
Signed-off-by: Álvaro Aguilar <alvaro.aguilar@scrm.lidl>